Address CI issues 2.

+++ code/pilotfile/plr.cpp
@@ -578,7 +578,7 @@ void pilotfile::plr_read_controls()
 			inv = handler->readInt("invert_axis");
 
 			if (idx < NUM_JOY_AXIS_ACTIONS) {
-				Control_config[idx + JOY_AXIS_BEGIN].take(CC_bind(CID_JOY0, axi, CCF_AXIS), 0);
+				Control_config[idx + JOY_AXIS_BEGIN].take(CC_bind(CID_JOY0, static_cast<short>(axi), CCF_AXIS), 0);
 				Control_config[idx + JOY_AXIS_BEGIN].invert(inv);
 			}
 		}
